created new pages,added Router, cards relocated from HomePage to gamePage, added Redirect to /404 if url is not exist
