--- .eslintcache
@@ -1,1 +1,0 @@
-[{"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\App.js":"1","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\index.js":"2","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\PokemonCard\\index.js":"3","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Footer\\index.js":"4","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Header\\index.js":"5","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Layout\\index.js":"6","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\routes\\Home\\index.js":"7","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\routes\\Game\\index.js":"8","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\Menu\\index.js":"9","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\Navbar\\index.js":"10","C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\index.js":"11"},{"size":591,"mtime":1612187280408,"results":"12","hashOfConfig":"13"},{"size":164,"mtime":1611746435887,"results":"14","hashOfConfig":"13"},{"size":2118,"mtime":1612169154210,"results":"15","hashOfConfig":"13"},{"size":264,"mtime":1611928083905,"results":"16","hashOfConfig":"13"},{"size":531,"mtime":1612178048820,"results":"17","hashOfConfig":"13"},{"size":695,"mtime":1612181848601,"results":"18","hashOfConfig":"13"},{"size":5844,"mtime":1612181942672,"results":"19","hashOfConfig":"13"},{"size":334,"mtime":1612185958322,"results":"20","hashOfConfig":"13"},{"size":1083,"mtime":1612190793647,"results":"21","hashOfConfig":"13"},{"size":623,"mtime":1612191759397,"results":"22","hashOfConfig":"13"},{"size":699,"mtime":1612190925658,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1q9e5jj",{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"41"},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\App.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\index.js",[],["48","49"],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\PokemonCard\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Footer\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Header\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\Layout\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\routes\\Home\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\routes\\Game\\index.js",[],["50","51"],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\Menu\\index.js",[],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\Navbar\\index.js",["52"],"C:\\Users\\compukter\\WebstormProjects\\pokemon-game\\src\\components\\MenuNavbar\\index.js",[],{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","replacedBy":"56"},{"ruleId":"53","replacedBy":"57"},{"ruleId":"55","replacedBy":"58"},{"ruleId":"59","severity":1,"message":"60","line":14,"column":17,"nodeType":"61","endLine":14,"endColumn":93},"no-native-reassign",["62"],"no-negated-in-lhs",["63"],["62"],["63"],"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
