+++ src/components/Header/index.js
@@ -1,12 +1,16 @@
 import s from './Header.module.css'
+import {useHistory} from 'react-router-dom'
 
-const Header =({title,descr,onClickButton}) =>{
+const Header =({title,descr}) =>{
+    let history = useHistory()
+
     let handleClick =() =>{
-        console.log('<Header/>')
-        onClickButton&&onClickButton()
+        history.push('/game')
     }
     return <header className={s.root}>
         <div className={s.forest}></div>
+        <div className={s.silhouette}></div>
+        <div className={s.moon}>  </div>
         <div className={s.container}>
             <h1>{title}</h1>
             <p>{descr}</p>
