+++ src/components/Layout/index.js
@@ -1,7 +1,7 @@
 import s from './Layout.module.css'
 
 
-const Layout =({id,title,urlBg,colorBg,children}) =>{
+const Layout =({id,title,urlBg,colorBg,children,}) =>{
     const background = {background:`url(${urlBg?urlBg:null}) left bottom fixed`,backgroundColor:`${colorBg}`}
     return <section className={s.root} id={id} style={background}>
         <div className={s.wrapper} >
