+++ src/components/MenuNavbar/Menu/index.js
@@ -1,34 +1,43 @@
 import cn from 'classnames'
 import s from './style.module.css'
+import {Link} from "react-router-dom";
 
-const Menu = ({isActive}) => {
+const Menu = ({isOpen,setActiveInactive}) => {
+    const handleClick = ()=>{
+        setActiveInactive()
+    }
+    const MENU = [{
+        title: 'HOME',
+        to: '/'
+    },
+        {
+            title: 'GAME',
+            to: '/game'
+        },
+        {
+            title: 'ABOUT',
+            to: '/about'
+        },
+        {
+            title: 'CONTACT',
+            to: '/contact'
+        }
+    ]
 
     return (
-         <div
-            className={cn(s.menuContainer, {[s.active]: isActive}, {[s.deactive]: !isActive})}>
+        <div
+            className={cn(s.menuContainer, {[s.active]: isOpen===true}, {[s.deactive]: isOpen===false})}>
             <div className={s.overlay}/>
             <div className={s.menuItems}>
                 <ul>
-                    <li>
-                        <a href="#welcome">
-                            HOME
-                        </a>
-                    </li>
-                    <li>
-                        <a href="#game">
-                            GAME
-                        </a>
-                    </li>
-                    <li>
-                        <a href="#about">
-                            ABOUT
-                        </a>
-                    </li>
-                    <li>
-                        <a href="#contact">
-                            CONTACT
-                        </a>
-                    </li>
+                    {MENU.map(({title, to}, index) => (
+                        <li key={index} onClick={handleClick}>
+                            <Link to={to}>
+                                {title}
+                            </Link>
+                        </li>
+                        ))
+                    }
                 </ul>
             </div>
         </div>
