+++ src/components/MenuNavbar/Menu/style.module.css
@@ -46,6 +46,8 @@
     justify-content: center;
     align-items: center;
     overflow: hidden;
+    opacity: 0;
+    z-index: -1;
 }
 
 .menuContainer::before,
