+++ src/components/PokemonCard/PokemonCard.module.css
@@ -30,7 +30,16 @@
 }
 
 .pokemonCard.active {
-    transform: rotateY(0deg);
+    animation: flipFront 1000ms 300ms forwards;
+}
+
+@keyframes flipFront {
+    from {
+        transform: rotateY(180deg);
+    }
+    to {
+        transform: rotateY(0deg);
+    }
 }
 
 .cardFront,
