+++ src/routes/AboutPage/index.js
@@ -1,0 +1,13 @@
+
+const AboutPage = () => {
+
+    return (
+
+        <div>
+            This is About Page!!!
+            <br/>
+        </div>
+    )
+}
+
+export default AboutPage;
