+++ src/routes/ContactPage/index.js
@@ -1,0 +1,14 @@
+
+const ContactPage = () => {
+
+    return (
+
+        <div>
+            This is ContactPage!!!
+            <br/>
+
+        </div>
+    )
+}
+
+export default ContactPage;
